fix: rename packages
* Base package renamed to `safeconcurrency` by adding a `package.go`,
  this only effects the display module name for the pkg.go.dev site.
* Moved implementations down from `api/`.
* Renamed `pool` to `workpool`.
