+++ .github/workflows/test.yml
@@ -28,4 +28,4 @@ jobs:
       - name: Test flakiness
         # Run the unit tests with a high count to check for flakiness
         # Yes, we really can run the tests 100000 times in just a few seconds
-        run: go test -count 100000 -failfast ./api/...
+        run: go test -count 100000 -failfast -skip '^Example' ./...
