+++ api/safeconcurrencyerrors/constant.go
@@ -1,10 +1,4 @@
-package results
-
-// Stop is a special error that can be returned from the [github.com/Izzette/go-safeconcurrency/api/types.TaskCallback]
-// to stop processing results.
-//
-//nolint:errname
-const Stop = constantError("stop")
+package safeconcurrencyerrors
 
 // constantError is a custom error type that can be used to create constant errors.
 type constantError string
