+++ api/safeconcurrencyerrors/errors.go
@@ -1,0 +1,7 @@
+package safeconcurrencyerrors
+
+// Stop is a special error that can be returned from the [github.com/Izzette/go-safeconcurrency/api/types.TaskCallback]
+// to stop processing results.
+//
+//nolint:errname
+const Stop = constantError("stop")
