+++ api/safeconcurrencyerrors/errors_test.go
@@ -1,4 +1,4 @@
-package results
+package safeconcurrencyerrors
 
 import (
 	"errors"
