+++ examples/sharedpool_test.go
@@ -4,16 +4,16 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/Izzette/go-safeconcurrency/api/pool"
+	"github.com/Izzette/go-safeconcurrency/workpool"
 )
 
 // Demonstrates how [types.Pool] can be used to create a worker pool that executes heterogeneous tasks concurrently.
 // Here we execute both [types.Task] ([IntTask] and [LogTask]) and [types.MultiResultTask] ([StringTask]).
-// The pool is created with no shared resource, no buffering, and a concurrency of 1 using [pool.NewPoolBuffered].
-// Tasks are submitted with [pool.Submit] and [pool.SubmitMultiResult].
+// The pool is created with no shared resource, no buffering, and a concurrency of 1 using [workpool.NewPoolBuffered].
+// Tasks are submitted with [workpool.Submit] and [workpool.SubmitMultiResult].
 func Example_sharedPool() {
 	// Create a new pool with no shared resource, no buffering, and a concurrency of 1.
-	sharedpool := pool.NewPoolBuffered[any](nil, 1, 0)
+	sharedpool := workpool.NewPoolBuffered[any](nil, 1, 0)
 	// Close the types.Pool.
 	// It's important to close the pool only after all tasks have been submitted.
 	// This will also wait for the pool to finish processing all tasks.
@@ -23,7 +23,7 @@ func Example_sharedPool() {
 	sharedpool.Start()
 
 	// Submit a task that returns an int and handle the result.
-	value, err := pool.Submit[any, int](context.Background(), sharedpool, &IntTask{42})
+	value, err := workpool.Submit[any, int](context.Background(), sharedpool, &IntTask{42})
 	if err != nil {
 		fmt.Printf("Error from IntTask: %v\n", err)
 	} else {
@@ -31,12 +31,12 @@ func Example_sharedPool() {
 	}
 
 	// This task doesn't return any result or error.
-	pool.Submit[any, any](context.Background(), sharedpool, &LogTask{"A message sent from the pool by LogTask"})
+	workpool.Submit[any, any](context.Background(), sharedpool, &LogTask{"A message sent from the pool by LogTask"})
 
 	// Tasks can also stream multiple results.
-	// StringTask implements types.MultiResultTask, submit it to the pool using pool.SubmitMultiResult, passing a callback
+	// StringTask implements types.MultiResultTask, submit it to the pool using workpool.SubmitMultiResult, passing a callback
 	// that will be invoked for each result.
-	err = pool.SubmitMultiResult[any, string](
+	err = workpool.SubmitMultiResult[any, string](
 		context.Background(), sharedpool, &StringTask{[]string{"hello", "world"}},
 		func(_ctx context.Context, value string) error {
 			// This function is called for each result published by the task.
@@ -52,7 +52,7 @@ func Example_sharedPool() {
 	// This task won't run because the context is already canceled.
 	ctx, cancelFunc := context.WithCancel(context.Background())
 	cancelFunc()
-	_, err = pool.SubmitMultiResultCollectAll[any, string](ctx, sharedpool, &StringTask{})
+	_, err = workpool.SubmitMultiResultCollectAll[any, string](ctx, sharedpool, &StringTask{})
 	if err != nil {
 		fmt.Printf("Error submitting StringTask: %v\n", err)
 	}
