+++ results/base.go
